Move MaRMAT into separate importable file for testing
